+++ tests/models/test_ngsi_v2_subscriptions.py
@@ -75,6 +75,13 @@ class TestSubscriptions(unittest.TestCase):
         with self.assertRaises(ValidationError):
             notification.mqtt = mqttCustom
 
+        # test onlyChangedAttrs-field
+        notification.onlyChangedAttrs = True
+        notification.onlyChangedAttrs = False
+        with self.assertRaises(ValidationError):
+            notification.onlyChangedAttrs = dict()
+
+
     @clean_test(fiware_service=settings.FIWARE_SERVICE,
                 fiware_servicepath=settings.FIWARE_SERVICEPATH,
                 cb_url=settings.CB_URL)
@@ -112,7 +119,6 @@ class TestSubscriptions(unittest.TestCase):
                 ]
             },
             "expires": "2030-04-05T14:00:00Z",
-            "throttling": 5
         }
 
         sub = Subscription.parse_obj(sub_dict)
@@ -133,6 +139,12 @@ class TestSubscriptions(unittest.TestCase):
 
             compare_dicts(sub.dict(exclude={'id'}),
                           sub_res.dict(exclude={'id'}))
+
+        # test validation of throttling
+        with self.assertRaises(ValidationError):
+            sub.throttling = -1
+        with self.assertRaises(ValidationError):
+            sub.throttling = 0.1
 
     def tearDown(self) -> None:
         """